Fix gcc sign conversion warnings in espeak-ng.c.

+++ src/libespeak-ng/speech.c
@@ -237,7 +237,7 @@ ESPEAK_NG_API espeak_ng_STATUS espeak_ng_InitializeOutput(espeak_ng_OUTPUT_MODE 
 	return ENS_OK;
 }
 
-int GetFileLength(const char *filename)
+size_t GetFileLength(const char *filename)
 {
 	struct stat statbuf;
